+++ snd-xrec.c
@@ -3130,7 +3130,7 @@ void finish_recording(snd_state *ss, recorder_info *rp)
 			    rp->total_output_frames * rp->out_chans * mus_data_format_to_bytes_per_sample(rp->out_format));
   snd_close(rp->output_file_descriptor, rp->output_file);
   rp->output_file_descriptor = -1;
-  duration = (Float)rp->total_output_frames / (Float)(rp->srate);
+  duration = (Float)((double)(rp->total_output_frames) / (Float)(rp->srate));
   /* 25-Jun-00: this used to divide by chans, but rp->total_output_frames is in terms of frames (it was named total_out_samps) */
   reflect_recorder_duration(duration);
   str = (char *)CALLOC(PRINT_BUFFER_SIZE, sizeof(char));
